created 'add position' form, validation, and a button to get to it from the main portfolio screen

+++ MonoStockPortfolio.Entities/Position.cs
@@ -9,6 +9,6 @@ namespace MonoStockPortfolio.Entities
         public string Ticker { get; set; }
         public decimal Shares { get; set; }
         public decimal PricePerShare { get; set; }
-        public int ContainingPortfolioID { get; set; }
+        public long ContainingPortfolioID { get; set; }
     }
 }
